refacto(pr)
