+++ src/Plateform.lua
@@ -1,8 +1,21 @@
+Plateform = {}
 
-function splitLine(line)
+local function createRectangle(x, y, w, h)
+    return {x = x, y = y, w = w, h = h}
+end
+
+function Plateform.create(x, y, scale, path)
+    local asset = love.graphics.newImage(path)
+    local rectangle = {}
+    rectangle.hitbox = createRectangle(x, y, asset:getWidth() * (scale / 100), asset:getHeight() * (scale / 100))
+    rectangle.scale = scale
+    rectangle.asset = asset
+    return rectangle
+end
+
+function Plateform.splitLine(line)
 	local values = {}
     local i = 1
-
 	for value in line:gmatch("[^ ]+") do
         if i < 4 then
             value = tonumber(value)
@@ -13,24 +26,21 @@ function splitLine(line)
 	return values
 end
 
-
-function loadPlateforms(path)
+function Plateform.load(path)
     local plateforms = {}
-    local i = 1
 
     for line in love.filesystem.lines(path) do
-        local values = splitLine(line)
+        local values = Plateform.splitLine(line)
         local x = values[1]
         local y = values[2]
         local scale = values[3]
         local asset = values[4]
-        plateforms[i] = createPlateform(x, y, scale, asset)
-        i = i + 1
+        plateforms[#plateforms + 1] = Plateform.create(x, y, scale, asset)
     end
     return plateforms
 end
 
-function drawPlateforms(plateforms)
+function Plateform.draw(plateforms)
     for i = 1, #plateforms do
         love.graphics.push()
         love.graphics.setColor(1, 1, 1)
